QAnything/requirements.txt
huggingface_hub==0.19.4
modelscope==1.13.0
Pillow==10.2.0
numpy==1.24.3
PyMuPDF==1.24.4
opencv-python-headless==4.9.0.80
easyocr==1.7.1
onnxruntime==1.17.1
torch==2.1.2
torchvision==0.16.2
transformers
vllm==0.3.3
openai==1.12.0
concurrent-log-handler==0.9.25
sentencepiece==0.1.99
tiktoken==0.6.0
sanic==23.6.0
sanic_ext==23.6.0
faiss-cpu==1.8.0
openpyxl==3.1.2
langchain==0.1.9
pypinyin==0.50.0
python-docx==1.1.0
unstructured==0.12.4
unstructured[pptx]==0.12.4
unstructured[md]==0.12.4
accelerate==0.21.0
networkx==3.2.1
faster_whisper==1.0.1
python-dotenv==1.0.1
duckduckgo-search==5.3.0b4
html2text==2024.2.26
mistune==3.0.2
xgboost==2.0.3
pdfplumber==0.11.0
PyPDF2==3.0.1
markdownify==0.12.1
datrie==0.8.2
hanziconv==0.3.2
shapely==2.0.4
pyclipper==1.3.0.post5